+++ org.eclipse.lsp4j.debug/src/main/xtend-gen/org/eclipse/lsp4j/debug/Scope.java
@@ -40,28 +40,30 @@ public class Scope {
   
   /**
    * The variables of this scope can be retrieved by passing the value of variablesReference to the
-   * VariablesRequest.
+   * VariablesRequest. The value should be less than or equal to 2147483647 (2^31 - 1).
    */
   @NonNull
-  private Long variablesReference;
+  private Integer variablesReference;
   
   /**
    * The number of named variables in this scope.
    * <p>
    * The client can use this optional information to present the variables in a paged UI and fetch them in chunks.
+   * The value should be less than or equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
-  private Long namedVariables;
+  private Integer namedVariables;
   
   /**
    * The number of indexed variables in this scope.
    * <p>
    * The client can use this optional information to present the variables in a paged UI and fetch them in chunks.
+   * The value should be less than or equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
-  private Long indexedVariables;
+  private Integer indexedVariables;
   
   /**
    * If true, the number of variables in this scope is large or expensive to retrieve.
@@ -145,19 +147,19 @@ public class Scope {
   
   /**
    * The variables of this scope can be retrieved by passing the value of variablesReference to the
-   * VariablesRequest.
+   * VariablesRequest. The value should be less than or equal to 2147483647 (2^31 - 1).
    */
   @Pure
   @NonNull
-  public Long getVariablesReference() {
+  public Integer getVariablesReference() {
     return this.variablesReference;
   }
   
   /**
    * The variables of this scope can be retrieved by passing the value of variablesReference to the
-   * VariablesRequest.
+   * VariablesRequest. The value should be less than or equal to 2147483647 (2^31 - 1).
    */
-  public void setVariablesReference(@NonNull final Long variablesReference) {
+  public void setVariablesReference(@NonNull final Integer variablesReference) {
     this.variablesReference = Preconditions.checkNotNull(variablesReference, "variablesReference");
   }
   
@@ -165,11 +167,12 @@ public class Scope {
    * The number of named variables in this scope.
    * <p>
    * The client can use this optional information to present the variables in a paged UI and fetch them in chunks.
+   * The value should be less than or equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
   @Pure
-  public Long getNamedVariables() {
+  public Integer getNamedVariables() {
     return this.namedVariables;
   }
   
@@ -177,10 +180,11 @@ public class Scope {
    * The number of named variables in this scope.
    * <p>
    * The client can use this optional information to present the variables in a paged UI and fetch them in chunks.
+   * The value should be less than or equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
-  public void setNamedVariables(final Long namedVariables) {
+  public void setNamedVariables(final Integer namedVariables) {
     this.namedVariables = namedVariables;
   }
   
@@ -188,11 +192,12 @@ public class Scope {
    * The number of indexed variables in this scope.
    * <p>
    * The client can use this optional information to present the variables in a paged UI and fetch them in chunks.
+   * The value should be less than or equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
   @Pure
-  public Long getIndexedVariables() {
+  public Integer getIndexedVariables() {
     return this.indexedVariables;
   }
   
@@ -200,10 +205,11 @@ public class Scope {
    * The number of indexed variables in this scope.
    * <p>
    * The client can use this optional information to present the variables in a paged UI and fetch them in chunks.
+   * The value should be less than or equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
-  public void setIndexedVariables(final Long indexedVariables) {
+  public void setIndexedVariables(final Integer indexedVariables) {
     this.indexedVariables = indexedVariables;
   }